dwm: Fix borderwin positioning in stack
Borders on top of stack-windows, starting from
second one, used to be placed a little bit
higher than they ought to be, because they
didn't take into accout that stack windows
on above do have borders that take some space.

It's not actually a big visual deal, but from
X server perspective, the border-window is
effectively overlapped by window above and it
may become a pretty hard-to-debug surprise if
we decide to use feature like BottomIf for
stack_mode property in XWindowChanges(3).

Fixes:	81c85415b29d9ff22f05e29717ff3e4decf4ef89