dwm,config: Rework client borders
Optimize the algorithm that was introduced before
and support unique type of border for every layout.

Prior to this patch, single-side borders for clients
used to be rendered via creating new window and
every time position or color for the border had to
change, this window was unmmap and a fresh one was
created, which is somewhat overkill and probably
expensive. Now in case some border-window attributes
have to be changed, change them in-place, instead of
unmapping the window and mapping a new one.

As a result:
    - Master windows have left-side borders.
    - Stack windows have top-side borders.
    - Floating windows have borders on each side.
    - Windows in monocle layout have bottom-side border.
    - Fullscreen windows have no borders whatsoever.

Fixes:	81c85415b29d9ff22f05e29717ff3e4decf4ef89